Hitting a breakpoint now ignores characters
Removed the ability to cause an overflow/segmentation fault when entering a
string into the popup window for values to jump to (hopefully).

Now, when hitting a breakpoint we just call wgetch to just take one character of
input to signify that the user has read it (perhaps this should change to be
only when the user presses enter?).