Some code re-arrangments
Moved around a few things after running with some new Debug flags (which might
be a tad excessive) and receiving some errors.

Also moved all equality checks around to have the variable to check on the
right (where possible), which will be the default from now on so as not to
accidentally use a single '=' on accident.